removed `mpi` class from hdnnpy.settings.defaults
with this commit, hdnnpy/dataset/* no longer needs `config.py`,
  which is user configuration file
temporarily moved code that depends on `config.py`
  from hdnnpy/utils.py to hdnnpy/others.py
  from hdnnpy/dataset/construct_datasets.py to hdnnpy/hdnnp.py
