+++ hdnnpy/hdnnp.py
@@ -10,6 +10,7 @@ import pickle
 import shutil
 import sys
 
+import ase.io
 import chainer
 from chainer.dataset import concat_examples
 import chainer.training.extensions as ext
@@ -24,8 +25,7 @@ from hdnnpy.chainer_extensions import (Evaluator,
                                        set_log_scale,
                                        )
 from hdnnpy.dataset import (DatasetGenerator,
-                            construct_test_datasets,
-                            construct_training_datasets,
+                            HDNNPDataset,
                             )
 from hdnnpy.format import (parse_poscars,
                            parse_xyz,
@@ -34,15 +34,17 @@ from hdnnpy.model import (HDNNP,
                           SingleNNP,
                           loss_func,
                           )
+from hdnnpy.others import (ChainerSafelyTerminate,
+                           assert_settings,
+                           dump_config,
+                           dump_lammps,
+                           dump_skopt_result,
+                           dump_training_result,
+                           )
 from hdnnpy.preproc import PREPROC
 from hdnnpy.settings import stg
 from hdnnpy.updater import HDUpdater
-from hdnnpy.utils import (ChainerSafelyTerminate,
-                          assert_settings,
-                          dump_config,
-                          dump_lammps,
-                          dump_skopt_result,
-                          dump_training_result,
+from hdnnpy.utils import (MPI,
                           mkdir,
                           pprint,
                           )
@@ -55,14 +57,14 @@ def main():
     if stg.args.mode == 'train':
         try:
             preproc = PREPROC[stg.dataset.preproc](stg.dataset.nfeature) \
-                    if stg.mpi.rank == 0 else PREPROC[None]()
+                    if MPI.rank == 0 else PREPROC[None]()
             if stg.args.is_resume:
                 preproc.load(stg.args.resume_dir.with_name('preproc.npz'))
             datasets, elements = construct_training_datasets(parse_xyz(stg.dataset.xyz_file), preproc)
             preproc.save(stg.file.out_dir/'preproc.npz')
             dataset = DatasetGenerator(*datasets).holdout(stg.dataset.ratio)
             masters, result = train(dataset, elements)
-            if stg.mpi.rank == 0:
+            if MPI.rank == 0:
                 chainer.serializers.save_npz(stg.file.out_dir/'masters.npz', masters)
                 dump_lammps(stg.file.out_dir/'lammps.nnp', preproc, masters)
                 dump_training_result(stg.file.out_dir/'result.yaml', result)
@@ -72,7 +74,7 @@ def main():
     elif stg.args.mode == 'param-search':
         try:
             seed = np.random.get_state()[1][0]
-            seed = stg.mpi.comm.bcast(seed, root=0)
+            seed = MPI.comm.bcast(seed, root=0)
             result = gp_minimize(objective_func, stg.skopt.space,
                                  n_random_starts=stg.skopt.init_num,
                                  n_calls=stg.skopt.max_num,
@@ -80,7 +82,7 @@ def main():
                                  random_state=seed,
                                  verbose=True,
                                  callback=stg.skopt.callback)
-            if stg.mpi.rank == 0:
+            if MPI.rank == 0:
                 dump_skopt_result(stg.file.out_dir/'skopt_result.csv', result)
                 dump_config(stg.file.out_dir/'best_config.py')
         finally:
@@ -138,7 +140,7 @@ def objective_func(**params):
         stdout = sys.stdout
         sys.stdout = devnull
         preproc = PREPROC[stg.dataset.preproc](stg.dataset.nfeature) \
-                if stg.mpi.rank == 0 else PREPROC[None]()
+                if MPI.rank == 0 else PREPROC[None]()
         datasets, elements = construct_training_datasets(parse_xyz(stg.dataset.xyz_file), preproc)
         for dataset in DatasetGenerator(*datasets).kfold(stg.skopt.kfold):
             _, result = train(dataset, elements)
@@ -147,13 +149,15 @@ def objective_func(**params):
     return sum(results) / stg.skopt.kfold
 
 
-def train(dataset, elements):
+def train(dataset, elements, comm=None):
+    if comm is None:
+        comm = chainermn.create_communicator('naive', MPI.comm)
     result = {'training_time': 0.0, 'observation': []}
 
     # model and optimizer
     masters = chainer.ChainList(*[SingleNNP(element) for element in elements])
     master_opt = chainer.optimizers.Adam(stg.model.init_lr)
-    master_opt = chainermn.create_multi_node_optimizer(master_opt, stg.mpi.chainer_comm)
+    master_opt = chainermn.create_multi_node_optimizer(master_opt, comm)
     master_opt.setup(masters)
     master_opt.add_hook(chainer.optimizer_hooks.Lasso(stg.model.l1_norm))
     master_opt.add_hook(chainer.optimizer_hooks.WeightDecay(stg.model.l2_norm))
@@ -162,16 +166,16 @@ def train(dataset, elements):
         tag = train.tag
 
         # iterators
-        train_iter = chainer.iterators.SerialIterator(train, stg.dataset.batch_size // stg.mpi.size,
+        train_iter = chainer.iterators.SerialIterator(train, stg.dataset.batch_size // MPI.size,
                                                       repeat=True, shuffle=True)
-        test_iter = chainer.iterators.SerialIterator(test, stg.dataset.batch_size // stg.mpi.size,
+        test_iter = chainer.iterators.SerialIterator(test, stg.dataset.batch_size // MPI.size,
                                                      repeat=False, shuffle=False)
 
         # model
         hdnnp = HDNNP(train.elemental_composition, loss_func)
         hdnnp.sync_param_with(masters)
         main_opt = chainer.Optimizer()
-        main_opt = chainermn.create_multi_node_optimizer(main_opt, stg.mpi.chainer_comm)
+        main_opt = chainermn.create_multi_node_optimizer(main_opt, comm)
         main_opt.setup(hdnnp)
 
         # triggers
@@ -190,8 +194,8 @@ def train(dataset, elements):
         trainer.extend(ext.ExponentialShift('alpha', 1 - stg.model.lr_decay,
                                             target=stg.model.final_lr, optimizer=master_opt))
         evaluator = Evaluator(iterator=test_iter, target=hdnnp)
-        trainer.extend(chainermn.create_multi_node_evaluator(evaluator, stg.mpi.chainer_comm))
-        if stg.args.mode == 'train' and stg.mpi.rank == 0:
+        trainer.extend(chainermn.create_multi_node_evaluator(evaluator, comm))
+        if stg.args.mode == 'train' and MPI.rank == 0:
             trainer.extend(ext.LogReport(log_name='training.log'))
             trainer.extend(ext.PrintReport(['epoch', 'iteration', 'main/RMSE', 'main/d_RMSE', 'main/tot_RMSE',
                                             'validation/main/RMSE', 'validation/main/d_RMSE',
@@ -211,8 +215,8 @@ def train(dataset, elements):
             chainer.serializers.load_npz(trainer_snapshot, trainer)
             result = pickle.loads(interim_result.read_bytes())
             # remove snapshot
-            stg.mpi.comm.Barrier()
-            if stg.mpi.rank == 0:
+            MPI.comm.Barrier()
+            if MPI.rank == 0:
                 trainer_snapshot.unlink()
                 interim_result.unlink()
             stg.args.is_resume = False
@@ -230,6 +234,91 @@ def predict(dataset, elements):
     hdnnp.sync_param_with(masters)
     energies, forces = hdnnp.predict(*concat_examples(dataset[:]))
     return energies.data, forces.data
+
+
+def construct_training_datasets(tag_xyz_map, preproc):
+    if 'all' in stg.dataset.tag:
+        included_tags = sorted(tag_xyz_map)
+    else:
+        included_tags = stg.dataset.tag
+
+    params = {
+        'Rc': stg.dataset.Rc,
+        'eta': stg.dataset.eta,
+        'Rs': stg.dataset.Rs,
+        'lambda_': stg.dataset.lambda_,
+        'zeta': stg.dataset.zeta,
+        }
+
+    datasets = []
+    elements = set()
+    stg.dataset.nsample = 0
+    for tag in included_tags:
+        try:
+            xyz_path = tag_xyz_map[tag]
+        except KeyError:
+            pprint(f'Sub dataset tagged as "{tag}" does not exist. Skipped.')
+            continue
+
+        pprint(f'Construct sub dataset tagged as "{tag}"')
+        dataset = HDNNPDataset(descriptor='symmetry_function',
+                               property_='interatomic_potential',
+                               order=1)
+        atoms = ase.io.read(xyz_path, index=':', format='xyz')
+
+        descriptor_npz_path = xyz_path.with_name('Symmetry_Function.npz')
+        if descriptor_npz_path.exists():
+            dataset.descriptor_dataset.load(descriptor_npz_path, verbose=True)
+        else:
+            dataset.descriptor_dataset.make(atoms, params, verbose=True)
+            dataset.descriptor_dataset.save(descriptor_npz_path, verbose=True)
+
+        property_npz_path = xyz_path.with_name('Interatomic_Potential.npz')
+        if property_npz_path.exists():
+            dataset.property_dataset.load(property_npz_path, verbose=True)
+        else:
+            dataset.property_dataset.make(atoms, verbose=True)
+            dataset.property_dataset.save(property_npz_path, verbose=True)
+
+        dataset.construct(preproc, shuffle=True)
+
+        dataset.scatter()
+        datasets.append(dataset)
+        elements.update(dataset.elements)
+        stg.dataset.nsample += dataset.total_size
+    pprint()
+    return datasets, elements
+
+
+def construct_test_datasets(tag_xyz_map, preproc):
+    params = {
+        'Rc': stg.dataset.Rc,
+        'eta': stg.dataset.eta,
+        'Rs': stg.dataset.Rs,
+        'lambda_': stg.dataset.lambda_,
+        'zeta': stg.dataset.zeta,
+        }
+
+    datasets = []
+    elements = set()
+    for xyz_path in tag_xyz_map.values():
+        dataset = HDNNPDataset(descriptor='symmetry_function',
+                               order=1)
+        atoms = ase.io.read(xyz_path, index=':', format='xyz')
+
+        descriptor_npz_path = xyz_path.with_name('Symmetry_Function.npz')
+        if descriptor_npz_path.exists():
+            dataset.descriptor_dataset.load(descriptor_npz_path, verbose=False)
+        else:
+            dataset.descriptor_dataset.make(atoms, params, verbose=False)
+            dataset.descriptor_dataset.save(descriptor_npz_path, verbose=False)
+
+        dataset.construct(preproc, shuffle=False)
+
+        datasets.append(dataset)
+        elements.update(dataset.elements)
+
+    return datasets, elements
 
 
 if __name__ == '__main__':
