+++ test/test.sh
@@ -6,13 +6,13 @@ timeout="gtimeout 5"
 mpirun="mpirun -np 2"
 hdnnpy="hdnnpy"
 
-# ${clean}; ${hdnnpy} train --verbose
-#
-# ${clean}; ${mpirun} ${hdnnpy} train --verbose
-#
-# ${clean}; ${hdnnpy} param-search
-#
-# ${clean}; ${mpirun} ${hdnnpy} param-search
+${clean}; ${hdnnpy} train --verbose
+
+${clean}; ${mpirun} ${hdnnpy} train --verbose
+
+${clean}; ${hdnnpy} param-search
+
+${clean}; ${mpirun} ${hdnnpy} param-search
 
 ${clean}; ${hdnnpy} sym-func
